scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudscraper>=1.2.71",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.3",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
]
